Resolve Design Limitation: Only one behavior was allowed per display
Now, a display can have multiple behaviors, and each behavior implements
Listener interface which means each behavior can be notified and a
corresponding action can be taken in the respective display.

What is still missing is the flow from Notify() being called to
a real method associated with a display being called, so that the
actual processing can happen.
  For instance, for the currentConditionsDisplay, the display needs
  to update its content after Notify is called on its Listener ( or
  behavior)